【完善】现在凸起的导航栏隐藏时会把凸起隐藏

+++ app/src/main/java/com/zjl/finalarchitecture/widget/navigationBar/CustomBottomNavigationView.java
@@ -90,7 +90,7 @@ public class CustomBottomNavigationView extends BottomNavigationView {
         return false;
     }
 
-    public float getFabDiameter() {
+    public int getFabDiameter() {
         return fabDiameter;
     }
